Namespaces for errors and agent metrics (supportability and CLR)
