+++ tests/Agent/IntegrationTests/IntegrationTests/AgentMetrics/RequiredSupportabilityMetrics.cs
@@ -10,7 +10,7 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests.AgentFeatures
+namespace NewRelic.Agent.IntegrationTests.AgentMetrics
 {
     // This test verifies that all supportability metrics are generated that are required by APM.
     [NetFrameworkTest]
